refactor(rib): split update-group payloads
